Revert "lesskey: add -F option"
This reverts commit 7e76db520c747ac3cb93eaeaef97bc8a9f69900e.

I often "mistakenly" hit q even if less quitted for the -F option. This
is because I cannot tell instantly whether less has quit or not. For
more predictable behavior, I'm reverting to the previous behavior.

I also found a bug of old less where nothing is printed when it
automatically quits.