ssh-stop: Directly specify socket name with ControlPath option
The previous implementation of ssh-stop relied upon a particular pattern
configured in the default ssh_config for naming control sockets. The new
implementation can handle any sockets in the ~/.ssh directory regardless
of the filename. This allows more flexible configuration for the
ControlPath option.

This commit adds ~/.ssh/ctrl-%C as an alternative value for the
ControlPath option. It can be useful when ~/.ssh/%r@%h:%p expands to a
filename that is too long for a Unix-domain socket name because of a
long host name.

Fixes https://github.com/magicant/settings/issues/16