yashrc-old: use new history file location if it exists
Improper installation of yash may fail to load the common initialization
file in yashrc, in which case the yashrc-old file is used instead. The
previous implementation of yashrc-old always used the historical
location of the history file, re-creating it even if the file has been
moved to the new location. This commit changes the behavior to use the
new location if it exists, and fall back to the old location otherwise.

This situation is often encountered when I casually invoke a development
build of yash in a working directory.