outdated: trap after opening .outdated.tmp
If trap actions are installed before opening the .outdated.tmp file, the
trap may unexpectedly remove the file even if the redirection fails
because of an existing file. To make sure the script removes the file
that has been created by the current script, trap should be configured
after the redirection.

This change makes it possible for a file created by the script to be
left un-removed if the script is interrupted just after the file
creation before installing the trap. Because there is no way to open a
file and set traps atomically, this behavior cannot be fully avoided.